requirements.txt
cryptography==2.8
robotframework==3.1.2
#not working in python38
paramiko==2.7.0
jira==2.0.0
slackclient==2.4.0
#not working in python38
#pycrypto>=2.6.1
pycryptodome==3.9.4
pywinrm==0.4.1
PyYAML==5.2
setuptools_scm==3.3.3
pytest-runner==5.2
pbr==5.4.4
pycparser>=2.19
cffi==1.13.2
psutil==5.6.7
#not working in python38
#psycopg2==2.8.4
psycopg2-binary
SQLAlchemy==1.3.11
tabulate==0.8.6
#for python3.x
beautifulsoup4==4.8.1
robotframework-selenium2library==3.0.0
sphinx==2.2.2
pytest==5.3.1
pylint==2.4.4
Pypubsub==3.3.0
#robotframework-ride==1.5.2.1
